Set up Resources.Load to replace CardSprites singleton
Instead of a SerializeField array of sprites to be assigned in the inspector to the CardSprite singleton, the sprites were renamed and moved to the resources folder  and are now called with Resources.Load based on the suit and rank names in  CardData. Updated CardData tests to test this new feature.